Converted to class based implementation and migrated example to testing folder
